рефакторинг отчет printRepPerson()- вывод строки и raschetTotalPay()- расчет заработной платы.

+++ src/SoftwareDevelopDomain/Settings.java
@@ -2,8 +2,8 @@ package SoftwareDevelopDomain;
 
 public class  Settings {
 
-   public static final byte WORKHOURSINMONTH = (byte) 160;
-   public static final byte WORKHOURSINDAY = 8;
+   public static final  int WORKHOURSINMONTH = 160;
+   public static final int WORKHOURSINDAY = 8;
     public  static final double PAYPERHOUR = 1000;
 
 }
